+++ lib/configs/colors.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+final Color bgColor = Color(0xFFF2FBFE);
+final Color white = Color(0xFFffffff);
+final Color curtain = Color(0x80ffffff);
+final Color shadow = Color(0xFF505659);
+final Color wavecolor = Color(0xFF97DDFA);
+final Color waveshadow = Color(0xFF83C1BB);
+final Color goalTextColor = Color(0xFF2870A0);
+final Color priceColor = Color(0xFF44AAD6);
+final Color textColor = Color(0xFF332F2E);
