+++ lib/main.dart
@@ -1,8 +1,9 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 
-import 'signup.dart';
-import 'login.dart';
+import 'views/signup.dart';
+import 'views/login.dart';
+import 'configs/colors.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -46,15 +47,6 @@ class StartPage extends StatefulWidget {
 }
 
 class _StartPageState extends State<StartPage> {
-  final Color bgColor = Color(0xFFF2FBFE);
-  final Color white = Color(0xFFffffff);
-  final Color curtain = Color(0x80ffffff);
-  final Color shadow = Color(0xFF505659);
-  final Color wavecolor = Color(0xFF97DDFA);
-  final Color waveshadow = Color(0xFF83C1BB);
-  final Color goalTextColor = Color(0xFF2870A0);
-  final Color priceColor = Color(0xFF44AAD6);
-  final Color textColor = Color(0xFF332F2E);
 
   @override
   Widget build(BuildContext context) {
